Mobile nav bar fix 2
When you added the closing </li> tag to the navbar elements, html interprets each new line to be a new space which is which is why I had no closing tag for them because it creates a small gap between the things on the navbar. This gap was extending the nav bar and which 'bugged' the @media queries I used. I fixed it in css with a font 0 workaround instead so the html doesn't look weird.